restructuring reducers
